python_agents/requirements.txt
# LangGraph Python Agents Requirements
# Core LangGraph and LangChain dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0

# LLM Providers
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-google-genai>=0.1.0

# Additional utilities
pydantic>=2.0.0
python-dotenv>=1.0.0
uuid>=1.30
typing-extensions>=4.0.0

# Optional: For enhanced functionality
httpx>=0.25.0
aiohttp>=3.8.0
